requirements.txt
django-tailwind
Pillow
django-tailwind[reload]
opencv-python
onnxruntime
numpy
django-widget-tweaks
djangorestframework
python-dotenv
# torch==2.6.0+cpu
# torchaudio==2.6.0+cpu
# torchvision==0.21.0+cpu
openvino